Minor improvements - configure git only if necessary and NERDTree remaps (#2)

+++ install.sh
@@ -1,14 +1,18 @@
+#!/usr/bin/env bash
+
 DOTFILES_ROOT=$(pwd -P)
 
-echo 'What is your git user name?'
-read -e git_author
+if ! test -f git/.gitconfig.local.symlink; then
+    echo 'What is your git user name?'
+    read -e git_author
 
-echo 'What is your git email?'
-read -e git_email
+    echo 'What is your git email?'
+    read -e git_email
 
-sed -e "s/AUTHOR/$git_author/g" -e "s/EMAIL/$git_email/g" git/.gitconfig.symlink > git/.gitconfig.local.symlink
+    sed -e "s/AUTHOR/$git_author/g" -e "s/EMAIL/$git_email/g" git/.gitconfig.symlink > git/.gitconfig.local.symlink
+fi
 
-ln $DOTFILES_ROOT/git/.gitconfig.symlink ~/.gitconfig
+ln $DOTFILES_ROOT/git/.gitconfig.local.symlink ~/.gitconfig
 
 ln $DOTFILES_ROOT/psql/.psqlrc.symlink ~/.psqlrc
